@@ -10,39 +10,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.12",
-    "@angular/common": "~13.3.12",
-    "@angular/compiler": "~13.3.12",
-    "@angular/core": "~13.3.12",
-    "@angular/forms": "~13.3.12",
-    "@angular/platform-browser": "~13.3.12",
-    "@angular/platform-browser-dynamic": "~13.3.12",
-    "@angular/router": "~13.3.12",
-    "@kolkov/ngx-gallery": "^1.2.4",
-    "bootstrap": "^4.5.3",
+    "@angular/animations": "^14.2.12",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
+    "@kolkov/ngx-gallery": "^2.0.1",
+    "@microsoft/signalr": "^7.0.0",
+    "bootstrap": "^5.2.3",
     "bootswatch": "^5.2.2",
     "font-awesome": "^4.7.0",
-    "ng2-file-upload": "^1.4.0",
-    "ngx-bootstrap": "^7.1.2",
-    "ngx-spinner": "^12.0.0",
+    "ng2-file-upload": "^2.0.0-3",
+    "ngx-bootstrap": "^9.0.0",
+    "ngx-spinner": "^14.0.0",
     "ngx-timeago": "^2.0.0",
-    "ngx-toastr": "^14.3.0",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.3.0",
+    "ngx-toastr": "^15.2.2",
+    "rxjs": "^7.5.7",
+    "tslib": "^2.4.1",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.10",
-    "@angular/cli": "~13.3.10",
-    "@angular/compiler-cli": "~13.3.12",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
     "@types/jasmine": "~3.8.0",
-    "@types/node": "^12.11.1",
+    "@types/jasminewd2": "^2.0.10",
+    "@types/node": "^18.11.9",
+    "codelyzer": "^6.0.2",
     "jasmine-core": "~3.8.0",
     "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
+    "karma-chrome-launcher": "^3.1.1",
     "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
     "typescript": "~4.6.4"
   }
 }
